Fix: Check for WP_Error `in evf_get_allowed_html_tags` (#813)
* Update evf-core-functions.php

The `wp_safe_remote_get` function returns an `array|WP_Error`. In the previous check, it was too optimistic which was causing errors on a few of the websites I manage. The error was `Uncaught Error: Cannot use object of type WP_Error as array` due to the check of the status code. Performing the `is_wp_error` alleviates this, as anything other than an `WP_Error` is a success in this context anyway.

* Remove whitepsace from empty line